+++ src/core_modules/capture-core/components/Pages/EnrollmentAddEvent/EnrollmentAddEventPage.container.js
@@ -1,7 +1,8 @@
 // @flow
-import React from 'react';
+import React, { useCallback } from 'react';
 // $FlowFixMe
-import { useSelector, shallowEqual } from 'react-redux';
+import { useDispatch, useSelector, shallowEqual } from 'react-redux';
+import { addEnrollmentEventPageActionTypes, navigateToEnrollmentPage } from './enrollmentAddEventPage.actions';
 import { useProgramInfo } from '../../../hooks/useProgramInfo';
 import { EnrollmentAddEventPageComponent } from './EnrollmentAddEventPage.component';
 import { useEnrollment } from '../common/EnrollmentOverviewDomain/useEnrollment';
@@ -11,7 +12,7 @@ import { getScopeInfo } from '../../../metaData';
 import { pageStatuses } from './EnrollmentAddEventPage.constants';
 
 export const EnrollmentAddEventPage = () => {
-    const { programId, stageId, teiId, enrollmentId, orgUnitId } = useSelector(
+    const { programId, stageId, orgUnitId, teiId, enrollmentId } = useSelector(
         ({
             router: {
                 location: { query },
@@ -20,11 +21,24 @@ export const EnrollmentAddEventPage = () => {
             programId: query.programId,
             stageId: query.stageId,
             orgUnitId: query.orgUnitId,
-            enrollmentId: query.enrollmentId,
             teiId: query.teiId,
+            enrollmentId: query.enrollmentId,
         }),
         shallowEqual,
     );
+
+    const dispatch = useDispatch();
+
+    const handleCancel = useCallback(() => {
+        dispatch(navigateToEnrollmentPage(programId, orgUnitId, teiId, enrollmentId));
+    }, [dispatch, programId, orgUnitId, teiId, enrollmentId]);
+
+    const handleSave = useCallback(() => {
+        dispatch(navigateToEnrollmentPage(programId, orgUnitId, teiId, enrollmentId));
+    }, [dispatch, programId, orgUnitId, teiId, enrollmentId]);
+
+    // TODO: Validate query params
+    // Ticket: https://jira.dhis2.org/browse/TECH-669
     const { program } = useProgramInfo(programId);
     const programStage = [...program.stages.values()].find(item => item.id === stageId);
     const enrollmentSite = useEnrollment(teiId).enrollment;
@@ -42,9 +56,14 @@ export const EnrollmentAddEventPage = () => {
     if (!programStage) {
         return <span>[program stage placeholder]</span>;
     }
+
+    // TODO: Get data from enrollment collection for the rules engine
+    // Ticket: https://jira.dhis2.org/browse/TECH-635
+
     return (
         <EnrollmentAddEventPageComponent
             programStage={programStage}
+            stageId={stageId}
             programId={programId}
             orgUnitId={orgUnitId}
             enrollmentsAsOptions={enrollmentsAsOptions}
@@ -52,6 +71,12 @@ export const EnrollmentAddEventPage = () => {
             trackedEntityName={trackedEntityName}
             enrollmentId={enrollmentId}
             pageStatus={pageStatus}
+            enrollmentId={enrollmentId}
+            teiId={teiId}
+            onSave={handleSave}
+            onSaveSuccessActionType={addEnrollmentEventPageActionTypes.EVENT_SAVE_SUCCESS}
+            onSaveErrorActionType={addEnrollmentEventPageActionTypes.EVENT_SAVE_ERROR}
+            onCancel={handleCancel}
         />
     );
 };