+++ src/core_modules/capture-core/components/Pages/EnrollmentAddEvent/EnrollmentAddEventPage.types.js
@@ -1,14 +1,21 @@
 // @flow
-import { ProgramStage } from '../../../metaData';
+import type { ProgramStage } from '../../../metaData';
+import type { ExternalSaveHandler } from '../../WidgetEnrollmentEventNew';
 
 export type Props = {|
     programStage: ProgramStage,
-    enrollmentId: string,
-    programId: string,
-    orgUnitId: string,
     trackedEntityName: string,
     teiDisplayName: string,
     enrollmentsAsOptions: Array<Object>,
     pageStatus: string,
+    programId: string,
+    stageId: string,
+    orgUnitId: string,
+    teiId: string,
+    enrollmentId: string,
+    onSave: ExternalSaveHandler,
+    onSaveSuccessActionType: string,
+    onSaveErrorActionType: string,
+    onCancel: () => void,
     ...CssClasses,
 |};