+++ src/core_modules/capture-core/components/WidgetAddEvent/index.js
@@ -1,1 +1,0 @@
-export { WidgetAddEvent } from './WidgetAddEvent.component';