+++ src/core_modules/capture-core/components/WidgetEnrollmentEventNew/Validated/useOrganisationUnit.js
@@ -1,0 +1,51 @@
+// @flow
+import { useMemo, useEffect, useState } from 'react';
+import { useDataQuery } from '@dhis2/app-runtime';
+import log from 'loglevel';
+import { errorCreator } from '../../../../capture-core-utils';
+
+export const useOrganisationUnit = (orgUnitId: string) => {
+    const [orgUnit, setOrgUnit] = useState();
+    const { error, loading, data, refetch, called } = useDataQuery(
+        useMemo(
+            () => ({
+                organisationUnits: {
+                    resource: 'organisationUnits',
+                    id: ({ variables: { orgUnitId: id } }) => id,
+                    params: {
+                        fields: ['displayName, code'],
+                    },
+                },
+            }),
+            [],
+        ), {
+            lazy: true,
+        },
+    );
+
+    useEffect(() => {
+        refetch({ variables: { orgUnitId } });
+    }, [refetch, orgUnitId]);
+
+    useEffect(() => {
+        if (error) {
+            log.error(errorCreator('could not retrieve organisation unit name')({ error }));
+        }
+    }, [error]);
+
+    useEffect(() => {
+        setOrgUnit(
+            (loading || !called || error) ?
+                undefined : {
+                    id: orgUnitId,
+                    name: data?.organisationUnits?.displayName,
+                    code: data?.organisationUnits?.code,
+                },
+        );
+    }, [orgUnitId, data, loading, called, error]);
+
+    return {
+        error,
+        orgUnit,
+    };
+};