+++ src/core_modules/capture-core/components/WidgetEnrollmentEventNew/Validated/validated.types.js
@@ -1,0 +1,32 @@
+// @flow
+import type { ProgramStage, RenderFoundation, TrackerProgram } from '../../../metaData';
+import { typeof addEventSaveTypes } from '../DataEntry/addEventSaveTypes';
+import type { OrgUnit, ExternalSaveHandler } from '../common.types';
+
+export type ContainerProps = {|
+    program: TrackerProgram,
+    stage: ProgramStage,
+    formFoundation: RenderFoundation,
+    teiId: string,
+    enrollmentId: string,
+    orgUnitId: string,
+    onSaveExternal?: ExternalSaveHandler,
+    onSaveSuccessActionType?: string,
+    onSaveErrorActionType?: string,
+    onCancel?: () => void,
+|};
+
+export type Props = {|
+    programName: string,
+    stage: ProgramStage,
+    formFoundation: RenderFoundation,
+    orgUnit: OrgUnit,
+    ready: boolean,
+    id: string,
+    itemId: string,
+    onSave: (saveType: $Keys<addEventSaveTypes>) => void,
+    onCancel: () => void,
+    formRef: (formInstance: any) => void,
+    dataEntryFieldRef: (instance: any, id: string) => void,
+    ...CssClasses,
+|};